Added model.py
