+++ rnd/autogpt_server/pyproject.toml
@@ -34,6 +34,7 @@ cx-freeze = { git = "https://github.com/ntindle/cx_Freeze.git", rev = "main", de
 poethepoet = "^0.26.1"
 httpx = "^0.27.0"
 pytest-watcher = "^0.4.2"
+requests = "^2.32.3"
 
 [build-system]
 requires = ["poetry-core"]
